requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
diff-match-patch==20241021
Django==4.2.23
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.13.2
